Database versions of the nihexporter data

These functions cache the data from the nihexporter database in a local database for use in examples and vignettes. Indexes are created to make joining tables on natural keys efficient.

Usage

nihexporter_sqlite(path = NULL)

nihexporter_postgres(dbname = "nihexporter", ...)

has_nihexporter(type = c("sqlite", "postgresql"), ...)

copy_nihexporter(src, ...)

Arguments

path

location of sqlite database file

dbname, ...

Arguments passed on to src_postgres()